requirements.txt
python-dotenv
flask
flask-cors
langchain
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
pinecone
